% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{Test function}
\usage{
myfun(BPPARAM = MulticoreParam(4L, RNGseed = 42L), verbose = TRUE)
}
\arguments{
\item{BPPARAM}{A BiocParallelParam object}

\item{verbose}{TRUE or FALSE}
}
\description{
Test function
}
